Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/coordinate-systems/outer-svg-intrinsic-size-002-expected.txt



PASS no sizing attributes set
FAIL specified width assert_equals: checking width. expected 100 but got 300
FAIL modified specified width assert_equals: checking width. expected 400 but got 300
FAIL specified width and height assert_equals: checking width. expected 400 but got 300
FAIL specified width and modified specified height assert_equals: checking width. expected 400 but got 300
FAIL specified height assert_equals: checking height. expected 200 but got 150
FAIL modified specified height assert_equals: checking height. expected 250 but got 150
PASS no specified sizing attrs (after setting & removing them)
FAIL set a 10x8 viewBox assert_equals: checking width. expected 500 but got 300
FAIL modified viewBox to 50x20 assert_equals: checking width. expected 500 but got 300
FAIL adding specified width, in presence of specified viewBox assert_equals: checking width. expected 100 but got 300
FAIL modifiying specified viewBox, in presence of specified width assert_equals: checking width. expected 100 but got 300
FAIL removing specified width, in presence of specified viewBox assert_equals: checking width. expected 500 but got 300
FAIL adding specified height, in presence of specified viewBox assert_equals: checking width. expected 80 but got 300
FAIL modifiying specified viewBox, in presence of specified height assert_equals: checking width. expected 50 but got 300
